airflow/requirements.txt
apache-airflow
tweepy
snowflake-connector-python
